backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
pyodbc==4.0.39
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0